pyproject.toml
[tool.poetry]
name = "mvideo_eldorado_parser"
version = "0.1.0"
description = ""
authors = ["taimaskhanov11 <81189712+taimaskhanov11@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = ">=3.10,<3.11"
loguru = "^0.6.0"
bs4 = "^0.0.1"
requests = "^2.27.1"
selenium = "^4.1.0"
webdriver-manager = "^3.5.2"
PyYAML = "^6.0"
aiogram = "^2.19"
aiohttp = "^3.8.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = "^22.1.0"
auto-py-to-exe = "^2.17.0"
isort = "^5.10.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
